API-12088 remove common objects from agent and adjust structures
