+++ src/agent/rtm.ts
@@ -1,34 +1,38 @@
+import { RTMAPI } from "../internal";
 import type {
-  ListChatsResponse,
-  ListChatParameters,
+  AgentForTransfer,
+  ChangePushNotificationsRequest,
+  CreateCustomerResponse,
+  CustomerParameters,
   EmptyResponse,
-  ListThreadsParameters,
-  ListThreadsResponse,
+  Event,
   GetChatResponse,
+  GetCustomerResponse,
   ListArchivesParameters,
   ListArchivesResponse,
-  StartChatParameters,
-  StartChatResponse,
-  ResumeChatParameters,
-  ResumeChatResponse,
-  TransferChatParameters,
-  SendEventResponse,
-  SendRichMessagePostbackParameters,
-  GetCustomerResponse,
+  ListChatParameters,
+  ListChatsResponse,
   ListCustomersParameters,
   ListCustomersResponse,
-  CustomerParameters,
-  CreateCustomerResponse,
-  MulticastRecipients,
-  AgentForTransfer,
-  LoginResponse,
-  ChangePushNotificationsRequest,
+  ListThreadsParameters,
+  ListThreadsResponse,
   LoginRequest,
+  LoginResponse,
+  MulticastRecipients,
+  Properties,
+  Push,
   Pushes,
+  ResumeChatParameters,
+  ResumeChatResponse,
+  RoutingStatus,
+  RTMAPIOptions,
+  SendEventResponse,
+  SendRichMessagePostbackParameters,
   SetRoutingStatusResponse,
+  StartChatParameters,
+  StartChatResponse,
+  TransferChatParameters,
 } from "./structures";
-import { Event, Properties, Push, RoutingStatus, RTMAPIOptions } from "../objects";
-import { RTMAPI } from "../internal/index";
 
 export default class RTM extends RTMAPI {
   constructor(options?: RTMAPIOptions) {
@@ -37,7 +41,7 @@ export default class RTM extends RTMAPI {
 
   /**
    * Allows to subscribe a handler for a given push. Returns function to unsubscribe.
-   * Note: multiple subscriptions for the same push are not allowed in sigle websocket connection.
+   * Note: multiple subscriptions for the same push are not allowed in single websocket connection.
    * @param push - push name to subscribe to
    * @param handler - function receiving push payload
    */
@@ -56,6 +60,7 @@ export default class RTM extends RTMAPI {
     }
     return this.send("login", loginData);
   }
+
   /**
    * Change the firebase push notifications properties.
    * @param change - properties to change
@@ -106,6 +111,7 @@ export default class RTM extends RTMAPI {
   async getChat(chat_id: string, thread_id?: string): Promise<GetChatResponse> {
     return this.send("get_chat", { chat_id, thread_id });
   }
+
   /**
    * It returns a list of the chats an Agent has access to. Together with a chat, the events of one thread from this chat are returned.
    *
@@ -196,6 +202,7 @@ export default class RTM extends RTMAPI {
       ignore_requester_presence,
     });
   }
+
   /**
    * Removes a user from chat. Removing customer user type is not allowed. It's always possible to remove the requester from the chat.
    * @param chat_id - chat to remove user from
@@ -214,8 +221,9 @@ export default class RTM extends RTMAPI {
       ignore_requester_presence,
     });
   }
+
   /**
-   * Sends an Event object. Use this method to send a message by specifing the Message event type in the request.
+   * Sends an Event object. Use this method to send a message by specifying the Message event type in the request.
    * It's possible to write to a chat without joining it. The user sending an event will be automatically added to the chat
    * with the present parameter set to false.
    * @param chat_id - chat to send event to
